chore: add sigint

+++ server/src/main.rs
@@ -1,6 +1,9 @@
 use std::os::unix::net::{UnixListener, UnixStream};
+use std::process::exit;
 use std::thread;
 use stream_message::{Message, Request, Response};
+
+use signal_hook::{consts::SIGINT, iterator::Signals};
 
 fn handle_client(mut stream: UnixStream) {
     loop {
@@ -13,9 +16,18 @@ fn handle_client(mut stream: UnixStream) {
 }
 
 fn main() {
-    let _ = std::fs::remove_file("./server.sock");
     let listener = UnixListener::bind("./server.sock").unwrap();
-    let mut threads = vec![];
+    let mut signals = Signals::new([SIGINT]).unwrap();
+    let handle = thread::spawn(move || {
+        for sig in signals.forever() {
+            if sig == SIGINT {
+                let _ = std::fs::remove_file("./server.sock");
+                break;
+            }
+        }
+        exit(0);
+    });
+    let mut threads = vec![handle];
     for stream in listener.incoming() {
         let Ok(stream) = stream else {
             break;
